Verif token + owner

+++ index.js
@@ -1,10 +1,15 @@
 const express = require('express');
 const app = express();
 
-const fs = require('fs');
 var path = require('path');
+const { getTasks, getTaskById, addTask, updateTask, deleteTask } = require('./controllers/tasksController');
+const { login, register, getUserProfile } = require('./usersAuth');
+
+var cors = require('cors');
+const jwtUtils = require('./utils/jwt.utils');
 
 const PORT = 5500;
+
 
 app.listen(
     PORT,
@@ -14,154 +19,73 @@ app.listen(
 )
 
 app.use(express.json()); //Le body des request renvoient .json
+
 //app.use("/static", express.static(path.resolve(__dirname, "public", "static")));
 app.use("/static", express.static(path.join(__dirname, 'public/static')));
 
+app.use(cors());
 
-function getTasks() {
-    return new Promise((resolve, reject) => {
-        fs.readFile("tasks.json", function(err, data) {
-            // Check for errors
-            //if (err) throw err;
-            if (err) {
-                console.log(err);
-                resolve({ error: true });
-            }
-            console.log("Data read");
+app.use(function(req, res, next) { //Midleware
+    console.log('Time:', Date.now());
+    next();
+});
 
-            resolve(JSON.parse(data));
-        });
-    });
-}
 
-function getTaskById(id) {
-    return new Promise((resolve, reject) => {
-        fs.readFile("tasks.json", function(err, data) {
-            // Check for errors
-            //if (err) throw err;
-            if (err) {
-                console.log(err);
-                resolve({ error: true });
-            }
-            console.log("Data read");
-
-            let tasks = JSON.parse(data);
-
-            let task = tasks.find((val) => val.id == id);
-
-            resolve(task);
-        });
-    });
-}
-
-function addTask(newTask) {
-    return new Promise(async(resolve, reject) => {
-        let tasks = await getTasks(); //On recup
-
-        if (tasks != null) {
-            let newID = tasks.sort((a, b) => { //Tri
-                return a.id - b.id;
-            })[tasks.length - 1].id + 1;
-
-            let finalTask = {
-                    ...newTask,
-                    id: newID
-                }
-                //console.log(finalTask);
-
-            tasks.push(finalTask);
-
-            fs.writeFile("tasks.json", JSON.stringify(tasks), err => {
-                if (err) {
-                    console.log(err);
-                    resolve({ error: true });
-                }
-
-                console.log("Done writing"); // Success
-
-                resolve(finalTask);
-            });
-        }
-    });
-}
-
-function updateTask(id, updatedTask) {
-    return new Promise(async(resolve, reject) => {
-        let tasks = await getTasks(); //On recup
-
-        if (tasks != null) {
-            const newTasks = tasks.map(obj => {
-                if (obj.id == id) {
-                    return {...updatedTask, id: obj.id };
-                }
-                return obj;
-            });
-
-            fs.writeFile("tasks.json", JSON.stringify(newTasks), err => {
-                if (err) {
-                    console.log(err);
-                    resolve({ error: true });
-                }
-                console.log("Done writing"); // Success
-
-                resolve(updatedTask);
-            });
-        }
-    });
-}
-
-function deleteTask(id) {
-    return new Promise(async(resolve, reject) => {
-        let tasks = await getTasks(); //On recup
-
-        if (tasks != null) {
-            let taskToDelete = await getTaskById(id);
-
-            const newTasks = tasks.filter(obj => obj.id != id);
-
-            fs.writeFile("tasks.json", JSON.stringify(newTasks), err => {
-                if (err) {
-                    console.log(err);
-                    resolve({ error: true });
-                }
-                console.log("Done writing"); // Success
-
-                resolve(taskToDelete);
-            });
-        }
-    });
-}
 
 app.get('/', (req, res) => {
     res.sendFile(`${__dirname}/public/`);
 })
 
-app.get("/tasks", async(req, res) => {
+app.get("/api/tasks", async(req, res) => {
     console.log("GET /tasks");
+    /**-------------------------------------------------------------------------------------AUTH------ */
+    var headerAuth = req.headers['authorization']; //Get the token
+    var userId = jwtUtils.getUserId(headerAuth); //Check if it exist and get the userId
 
-    const tasks = await getTasks();
+    if (userId < 0) return res.status(400).json({ 'error': 'wrong token' });
+    /**----------------------------------------------------------------------------------------------- */
+
+    let tasks = await getTasks();
+
+    tasks = tasks.filter((task) => task.userId == userId); //ON VERIFIE LE PROPRIETAIRE !!!
+
     console.log(tasks);
 
     res.status(200).send(tasks);
     console.log("OK");
-
 })
 
-app.get("/tasks/:id", async(req, res) => {
+app.get("/api/tasks/:id", async(req, res) => {
     console.log("GET /tasks/:id");
+
+    /**-------------------------------------------------------------------------------------AUTH------ */
+    var headerAuth = req.headers['authorization']; //Get the token
+    var userId = jwtUtils.getUserId(headerAuth); //Check if it exist and get the userId
+
+    if (userId < 0) return res.status(400).json({ 'error': 'wrong token' });
+    /**----------------------------------------------------------------------------------------------- */
 
     const { id } = req.params;
 
-    const task = await getTaskById(id);
+    let task = await getTaskById(id);
+
+    if (task.userId != userId) return res.status(200).send({});
+
     console.log(task);
 
     res.status(200).send(task);
     console.log("OK");
-
 })
 
-app.post('/task', async(req, res) => {
+app.post('/api/task', async(req, res) => {
     console.log("POST /task");
+
+    /**-------------------------------------------------------------------------------------AUTH------ */
+    var headerAuth = req.headers['authorization']; //Get the token
+    var userId = jwtUtils.getUserId(headerAuth); //Check if it exist and get the userId
+
+    if (userId < 0) return res.status(400).json({ 'error': 'wrong token' });
+    /**----------------------------------------------------------------------------------------------- */
 
     const { name } = req.body;
     const { priority } = req.body;
@@ -172,6 +96,7 @@ app.post('/task', async(req, res) => {
 
     let newTask = {
         id: -1,
+        userId: userId,
         name: name,
         priority: priority
     }
@@ -182,15 +107,28 @@ app.post('/task', async(req, res) => {
     res.status(201).send(taskToSend);
 })
 
-app.put("/tasks/:id", async(req, res) => {
+app.put("/api/tasks/:id", async(req, res) => {
     console.log("PUT /tasks/:id");
+
+    /**-------------------------------------------------------------------------------------AUTH------ */
+    var headerAuth = req.headers['authorization']; //Get the token
+    var userId = jwtUtils.getUserId(headerAuth); //Check if it exist and get the userId
+
+    if (userId < 0) return res.status(400).json({ 'error': 'wrong token' });
+
+    /**----------------------------------------------------------------------------------------------- */
 
     const { id } = req.params;
     const { name } = req.body;
     const { priority } = req.body;
 
+    let taskToUpdate = await getTaskById(id);
+
+    if (taskToUpdate.userId != userId) return res.status(200).send({});
+
     let newTask = {
         id: parseInt(id),
+        userId: userId,
         name: name,
         priority: priority
     }
@@ -202,14 +140,48 @@ app.put("/tasks/:id", async(req, res) => {
     console.log("OK");
 })
 
-app.delete("/tasks/:id", async(req, res) => {
+app.delete("/api/tasks/:id", async(req, res) => {
     console.log("DELETE /tasks/:id");
 
+    /**-------------------------------------------------------------------------------------AUTH------ */
+    var headerAuth = req.headers['authorization']; //Get the token
+    var userId = jwtUtils.getUserId(headerAuth); //Check if it exist and get the userId
+
+    if (userId < 0) return res.status(400).json({ 'error': 'wrong token' });
+
+    /**----------------------------------------------------------------------------------------------- */
+
     const { id } = req.params;
+
+    let taskToDelete = await getTaskById(id);
+
+    if (taskToDelete.userId != userId) return res.status(200).send({});
 
     const task = await deleteTask(id);
     console.log(task);
 
     res.status(200).send(task);
     console.log("OK");
+})
+
+/** -------------------------------Authentification------------------------------------------ */
+
+app.post("/api/register", async(req, res) => {
+    console.log("POST /api/register");
+
+    register(req, res);
+    //res.status(201).send("Register process...");
+})
+
+app.post("/api/login", async(req, res) => {
+    console.log("POST /api/login");
+
+    login(req, res);
+    //res.status(201).send("Login process...");
+})
+
+app.get("/api/test", async(req, res) => {
+    console.log("TEST /api/test");
+
+    getUserProfile(req, res);
 })
